refactor: extract shared MIME type detection utility
